% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\alias{run_bayesnec_app}
\title{Run Shiny bayesnec Application}
\usage{
run_app()

run_bayesnec_app()
}
\description{
Run Shiny bayesnec Application
}
\section{Functions}{
\itemize{
\item \code{run_bayesnec_app}: 
}}

