% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{shinybayesnec-package}
\alias{shinybayesnec}
\alias{shinybayesnec-package}
\title{shinybayesnec: bayesnec Shiny App}
\description{
Shiny application for concentration response models and estimating no effect concentrations using bayesnec.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/open-aims/shinybayesnec}
  \item Report bugs at \url{https://github.com/open-aims/shinybayesnec/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au} (\href{https://orcid.org/0000-0001-5148-6731}{ORCID}) [contributor]

Authors:
\itemize{
  \item Gerard Ricardo \email{g.ricardo@uq.edu.au} (\href{https://orcid.org/0000-0002-7761-0806}{ORCID})
}

Other contributors:
\itemize{
  \item Diego Barneche \email{d.barneche@aims.gov.au} (\href{https://orcid.org/0000-0002-4568-2362}{ORCID}) [contributor]
}

}
\keyword{internal}
